Refactor class paths, addressing some differences in dev/prod
